fix some retry logic and speaker volume logic

+++ indra/newview/llvoicewebrtc.h
@@ -829,7 +829,7 @@ private:
 	
 	bool		mSpeakerVolumeDirty;
 	bool		mSpeakerMuteDirty;
-	int			mSpeakerVolume;
+	float		mSpeakerVolume;
 
 	int			mMicVolume;
 	bool		mMicVolumeDirty;
